refactor: convert Audit model to async
