+++ apps/meteor/ee/client/views/audit/hooks/useAuditTab.ts
@@ -1,9 +1,8 @@
+import type { IAuditLog } from '@rocket.chat/core-typings';
 import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
 import { useRoute, useRouteParameter } from '@rocket.chat/ui-contexts';
 import type { SetStateAction } from 'react';
 import { useMemo } from 'react';
-
-import type { IAuditLog } from '../../../../definition/IAuditLog';
 
 const typeToTabMap: Record<IAuditLog['fields']['type'], string> = {
 	'': 'rooms',
