+++ apps/meteor/ee/server/lib/audit/methods.ts
@@ -3,15 +3,13 @@ import { check } from 'meteor/check';
 import { DDPRateLimiter } from 'meteor/ddp-rate-limiter';
 import { TAPi18n } from 'meteor/rocketchat:tap-i18n';
 import { escapeRegExp } from '@rocket.chat/string-helpers';
-import type { ILivechatAgent, ILivechatVisitor, IMessage, IRoom, IUser } from '@rocket.chat/core-typings';
+import type { ILivechatAgent, ILivechatVisitor, IMessage, IRoom, IUser, IAuditLog } from '@rocket.chat/core-typings';
 import type { ServerMethods } from '@rocket.chat/ui-contexts';
 import type { Filter } from 'mongodb';
-import { LivechatRooms, Messages, Rooms, Users } from '@rocket.chat/models';
+import { LivechatRooms, Messages, Rooms, Users, AuditLog } from '@rocket.chat/models';
 
-import AuditLog from './AuditLog';
 import { hasPermissionAsync } from '../../../../app/authorization/server/functions/hasPermission';
 import { updateCounter } from '../../../../app/statistics/server';
-import type { IAuditLog } from '../../../definition/IAuditLog';
 import { isTruthy } from '../../../../lib/isTruthy';
 
 const getValue = (room: IRoom | null) => room && { rids: [room._id], name: room.name };
@@ -71,7 +69,7 @@ declare module '@rocket.chat/ui-contexts' {
 			endDate: Date;
 			users: NonNullable<IUser['username']>[];
 			msg: IMessage['msg'];
-			type: 'l';
+			type: string;
 			visitor?: ILivechatVisitor['_id'];
 			agent?: ILivechatAgent['_id'];
 		}) => IMessage[];
@@ -110,7 +108,7 @@ Meteor.methods<ServerMethods>({
 
 		// Once the filter is applied, messages will be shown and a log containing all filters will be saved for further auditing.
 
-		AuditLog.insert({
+		await AuditLog.insertOne({
 			ts: new Date(),
 			results: messages.length,
 			u: user,
@@ -161,7 +159,7 @@ Meteor.methods<ServerMethods>({
 
 		// Once the filter is applied, messages will be shown and a log containing all filters will be saved for further auditing.
 
-		AuditLog.insert({
+		await AuditLog.insertOne({
 			ts: new Date(),
 			results: messages.length,
 			u: user,
@@ -184,7 +182,7 @@ Meteor.methods<ServerMethods>({
 				$gt: startDate,
 				$lt: endDate,
 			},
-		}).fetch();
+		}).toArray();
 	},
 });
