+++ apps/meteor/ee/server/models/startup.ts
@@ -5,6 +5,7 @@ import { onLicense } from '../../app/license/server/license';
 // No functionality will be imported tho, just the service registration
 import('./LivechatPriority');
 import('./OmnichannelServiceLevelAgreements');
+import('./AuditLog');
 
 onLicense('livechat-enterprise', () => {
 	import('./CannedResponse');
