+++ packages/core-typings/src/index.ts
@@ -41,6 +41,7 @@ export * from './IUserStatus';
 export * from './IUser';
 
 export * from './ee/ILicense/ILicense';
+export * from './ee/IAuditLog';
 
 export * from './import';
 export * from './IIncomingMessage';
