+++ packages/model-typings/src/index.ts
@@ -71,3 +71,4 @@ export * from './models/IImportsModel';
 export * from './models/IRawImportsModel';
 export * from './models/IFederationRoomEventsModel';
 export * from './models/IAppsTokensModel';
+export * from './models/IAuditLogModel';
