+++ packages/model-typings/src/models/IAuditLogModel.ts
@@ -1,0 +1,5 @@
+import type { IAuditLog } from '@rocket.chat/core-typings';
+
+import type { IBaseModel } from './IBaseModel';
+
+export type IAuditLogModel = IBaseModel<IAuditLog>;
