+++ packages/models/src/index.ts
@@ -71,6 +71,7 @@ import type {
 	IRawImportsModel,
 	IFederationRoomEventsModel,
 	IAppsTokensModel,
+	IAuditLogModel,
 } from '@rocket.chat/model-typings';
 
 import { proxify } from './proxify';
@@ -156,3 +157,4 @@ export const MatrixBridgedUser = proxify<IMatrixBridgedUserModel>('IMatrixBridge
 export const OmnichannelServiceLevelAgreements = proxify<IOmnichannelServiceLevelAgreementsModel>(
 	'IOmnichannelServiceLevelAgreementsModel',
 );
+export const AuditLog = proxify<IAuditLogModel>('IAuditLogModel');
